move Leaves code into module
